pyproject.toml
[tool.poetry]
name = "django-school-api"
version = "0.1.0"
description = ""
authors = ["Mohammed-Salama <mohammedsalama1617@gmail.com>"]
readme = "README.md"
packages = [{include = "django_school_api"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.2"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
jwt = "^1.3.1"
pyjwt = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
